+++ src/nextjs/nextjs-wizard.ts
@@ -346,7 +346,7 @@ async function createOrMergeNextJsFiles(
       id: 'replay',
       prompt: `Do you want to enable ${chalk.bold(
         'Sentry Session Replay',
-      )} to get reproduction of frontend errors via user sessions?`,
+      )} to get a video-like reproduction of errors during a user session?`,
       enabledHint: 'recommended, but increases bundle size',
     },
   ] as const);